requirements.txt
python-telegram-bot==20.6
Flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0